Création du composant EligibilityResult pour factoriser le code
