+++ app/aides/pret-taux-0/pret-avance-renovation/PAR.tsx
@@ -1,10 +1,9 @@
 'use client'
 
-import { Card, CTA, CTAWrapper } from '@/components/UI'
+import { Card } from '@/components/UI'
 import rules from '@/app/règles/rules'
 import checkIcon from '@/public/check.svg'
-import investissementIcon from '@/public/investissement.svg'
-import Publicodes, { formatValue } from 'publicodes'
+import Publicodes from 'publicodes'
 import parImage from '@/public/par.png'
 import Image from 'next/image'
 import {
@@ -24,13 +23,10 @@ import {
   getSituation,
 } from '@/components/publicodes/situationUtils'
 import { useSearchParams } from 'next/navigation'
-import { useMediaQuery } from 'usehooks-ts'
-import AmpleurCTA from '@/app/module/AmpleurCTA'
-import { Key } from '@/components/explications/ExplicationUI'
 import FatConseiller from '@/components/FatConseiller'
+import { EligibilityResult } from '@/components/EligibilityResult'
 
 export default function PAR() {
-  const isMobile = useMediaQuery('(max-width: 400px)')
   const engine = new Publicodes(rules)
   const dottedName = 'PAR'
   const setSearchParams = useSetSearchParams()
@@ -154,76 +150,15 @@ export default function PAR() {
           </Li>
         </QuestionList>
         {!Object.keys(evaluation.missingVariables).length && (
-          <div
-            css={`
-              background: var(--lightestColor);
-              border-bottom: 4px solid var(--color);
-              padding: 1rem;
-              display: flex;
-              gap: 1rem;
-              justify-content: space-between;
-              align-items: center;
-              text-align: center;
-              flex-wrap: wrap;
-            `}
-          >
-            {!isMobile && (
-              <Image src={investissementIcon} alt="icone montant en euro" />
-            )}
-            <p
-              css={`
-                flex: 1;
-              `}
-            >
-              {evaluation.nodeValue ? (
-                <>
-                  <strong>Vous êtes éligible à un prêt d'un montant</strong>
-                  <span
-                    css={`
-                      display: block;
-                      margin: 0.5rem 0;
-                    `}
-                  >
-                    de{' '}
-                    <Key
-                      $state="prime"
-                      css={`
-                        font-size: 1.5rem;
-                        padding: 0.4rem;
-                      `}
-                    >
-                      {formatValue(engine.evaluate(dottedName + ' . montant'))}
-                    </Key>{' '}
-                    maximum
-                  </span>
-                  <small>
-                    sans intérêt pendant{' '}
-                    <Key $state="in-progress">
-                      {formatValue(engine.evaluate(dottedName + ' . durée'))}
-                    </Key>
-                  </small>
-                </>
-              ) : (
-                <>
-                  <span
-                    css={`
-                      color: red;
-                    `}
-                  >
-                    Vous n'êtes pas éligible au Prêt avance mutation
-                  </span>
-                  <br />
-                  <span>⚠️ Vous êtes peut-être éligible à d'autres aides!</span>
-                  <br />
-                </>
-              )}
-            </p>
-            <CTAWrapper $justify="left" $customCss="margin: 0;">
-              <CTA $importance="primary" css="font-size: 100%">
-                <AmpleurCTA {...{ situation: situation }} />
-              </CTA>
-            </CTAWrapper>
-          </div>
+          <EligibilityResult
+            {...{
+              evaluation,
+              engine,
+              dottedName,
+              situation,
+              text: 'au PAR+',
+            }}
+          />
         )}
       </Card>
       <h3>Comment cela fonctionne?</h3>
